ZO-6370: move image_parallax_properties to image_row
